requirements.txt
streamlit
PyPDF2
dotenv
google-generativeai